spring-cache-memcached/src/main/java/com/isa/spring/cache/memcached/config/CacheProperties.java
package com.isa.spring.cache.memcached.config;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

import java.util.Arrays;
import java.util.Collections;
import java.util.List;

@Component
public class CacheProperties {
    @Value("${cache.enabled}")
    private boolean enabled;

    @Value("${cache.name}")
    private String name;

    @Value("${cache.timeout}")
    private int timeout;

    @Value("${cache.default-ttl}")
    private int defaultTtl;

    @Value("${cache.servers}")
    private String servers;

    public boolean isEnabled() {
        return enabled;
    }

    public void setEnabled(boolean enabled) {
        this.enabled = enabled;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public int getTimeout() {
        return timeout;
    }

    public void setTimeout(int timeout) {
        this.timeout = timeout;
    }

    public int getDefaultTtl() {
        return defaultTtl;
    }

    public void setDefaultTtl(int defaultTtl) {
        this.defaultTtl = defaultTtl;
    }

    public String getServers() {
        return servers;
    }

    public void setServers(String servers) {
        this.servers = servers;
    }
}
